implemented the sketch.py according to the feedback
